Chi^2 Min of Arrays
Added Ifit.C from Root's fitting examples to see how to fit arrays of data instead of histograms and graphs. This method is necessary once I use data sets for different energies. Test_Min.C is just modifying Ifit.C to fit a few quadratic data points and seems to be working. Next up is making a XS function that reads in energy and angle and minimizing that to different data sets. Hopefully this will help with the charge and magnetic FF separation. If not I'll try fixing Fm and calculating Fch and then interating back and forth between the two.