Working on a Minimizer for World Data
Minuit_Test. C was to make Minuit work with my XS equation and Amroun's data. Global_Fit_3He_SOG.C is the code being developed to fit the 3He world data. Currently the minimizer seems to run on the data, but I need to extract and print the FFs. Also realized in SOG_XS_Fit.C I had the wrong E0. It is now 0.64 GeV and was 0.5084 GeV.