3He Minimizer FF Plots and Chi^2 Results
The Minimizer code has been updated to extract the FFs from the fit parameters and then plot Fch and Fm. Also added the spreadsheet of world data which still needs a lot of work. Fits are working ok, but there are usually a few high Chi^2 points messing with the fits. Amroun used more data than in the 1994 paper for the fits so I will work to add that data to my data set before fitting to see if that improves the FF extractions.