Updated Random Radii Generator
Updated the Ri generator so that it generates increments of 0.1. Also attempting to get a makefile to work so I can build an executable to speed up the code. Not sure why I'm having issues with the makefile. Also now print the fit results to a text file and can loop the code for many different generted radii.