Recovered/Added Arnold's Data
Amroun 1994 ref 7 Arnold 1978 gave its FFs in a strange way using A^1/2. I was able to work it backward to get real cross sections back out. Need to double check that I propagated the errors correctly. This is high Q^2 data but with large error bars. Still it seems to help the fit a bit. Updated input text file as well. Added color coding for the Arnold data set to the plots. Also tried some calculations in the data spreadsheet to see if I could recover any of the data from the other papers which don't list initial energy or angle but was unsuccessful.