Automated Fine Tuning Ri Values
Now if I implement reasonable Ri guesses for a given number of Gaussians the code will adjust each Ri up and down in 0.1 fm increments. It will then stop when chi^2 stops improving and store the best value. This gives a brute force way to "optimize" the Ri values. So far implemented for ngaus=8,10. Both don't seem to give as good of a fit (mostly for physical reasons like radii and FFs) as ngaus=12.