Attempting to Compile Fits as Executable
Currently Test.C is attempting to mirror Global_Fit_3He_SOG.C. I have cleaned the file so that it uses slightly better coding practices than previously. It can now be compiled as an executable and the first test of the so far best n=12 single fit took 0.4 minutes compared to 2.5 min previously so after more rigorous testing I'm hoping for a run speed increase of about 5X. The current executable doesn't pop up the canvases with plots so I may need to save the pics like with the VDC code or figure out how to still get plots to pop out.