Added Bootstrap Still Working on Visualization
Added the ability to bootstrap the same set of Ri values by randomly selecting a number of points equal to the size of the dataset from the dataset allowing repeats. This output prints to a text file to be read for plotting. Multifit_FF_Plots.C plots the resulting bootstrap fits. However, it still doesn't seem to print them all correctly on the same plot. Need to fix that next.