Finally Have All Curves Plotting
Turns out to get all of the curves to plot with a user defined function in root you do need to use parameters in the function instead of just variables. No idea why the previous attempts failed still, but this new method is working for the mag FF. I will fix the charge FF and make the plots pretty again when I get back from the gym. Also make the code smarter with a loop for setting parameters and accounting for ngaus. Test_Multi_Plot.C was simply to try out multicurve plotting. Partial victory!