Added Charge Densities for FB
Added charge densities for FB fits as well as charge radius calculations. Also tried to fit the FB XS fit function, but it still seems pretty crappy. Fixed that GE and GEM are really Fch and Fm in the variables used. Also added cuts for sum Qim in the multiplot code.