Updated Multifit Plot for Charge Radii
The Multifit_FF_Plots.C code now calculates the rms charge radii as -6 * the derivative of the charge FF at Q^2=0. It also plots a histogram of the results and fits a Gaussian to the result to give statistics. Still somewhat inelegantly implements. Need to be careful when changing number of Gaussians.