Added Qi Plots to Multifit and XS Plot to Plot_FFs
I added plots of the individual Qi distributions to the Multifit code. These distributions are then fit, mostly successfully, by Gaussians to determine their mean values. Attempts at using a Poisson failed. I then pulled the Ri and Qi average values out and plotted those as a test of some kind of grand final FFs. This was not very successful. The FF Shapes were OK, but the minima were pretty off. 

Also in Plot_FFs I used the FFs to plot the real XS at a given energy and angle against Q^2. It can be compared to Amroun's XS results. I need to pick a good representative FF fit and plug it in to make sure it seems reasonable since the average Ri and Qi did not.