Finished Sorting Algorithm for Repeat Fits
Decided to see how many Chi^2 values (and thus Ri combinations) were repeats out of the final fits. Code now finds the number of times a combination repeated. It also finds the top ten most repeated fits and allows a Chi^2 cutoff to be applied to this as well. Hopefully I can find a representative best fit out of these.