Added Ability to Plot Representative Fit
Repeat analysis now finds the the most repeated fits, their chi^2, and the first number of that fit that occurs. Multifit can now accept the number of a fit from the data file and draw that fit in a different color on the plot of all fits for comparison. Hopefully I can use this to choose a good representative fit to list as my parametrization.