Getting FF World Data to Plot with My FF Fits
Started this is Global Fit code but moved it to Plott_FFs for efficiency. Not hard now, but a pain and there's lots of data I could add. Probably need to do Fch and Fm separately and each experimental data set must be graphed separately to use different colored points.