Added More Options and Error Bars
Added more options for drawing the various functions and uncertainty bands in Multifit. Also made a dummy function to make a solid red line in the legend for my new fits. Also made the axis labels and titles larger in the Ri and Qi plots, but they're still not great. In Plot_FF  I changed my point from a marker to a real tgraph with error bars. I turned Amround XS back on for comparison. Also added a zoomed in nonlog plot to show my point in relation to the fits.