Added Comparisons Between Nuclei/Fixed 3H Rho
Added the Compare_FFs.C code to plot the 3H and 3He FFs and rhos against one another. Minor fixes in Plot and Rosenbluth. Fixed Multifit so that the Z value changes based on 3H or 3He. This was causing 3H rho to be twice as large previously (updated in thesis to be correct now).